Add information how to write good commit messages
Writing good commit messages with title and description is important
part of working with version control and should be included in level 1
competency for github. Developer should be aware of constraints to
length of commit message subject and body. He should also be able to
configure his ide to check commit message line lengths.